Add files via upload
advanced android emojify me (with additional library - butterknife)

Butterknife: If you've created an app with a lot of views, then you'll know how tedious it can be to write findViewById() calls for each one, as well as to create custom onClicks for each one. These tasks pertain to Data Binding, the realm of connecting your data to the user facing views. Butterknife is a revolutionary library that uses annotations to make this much easier.